refactor: cleanup and restructure files
