+++ where_to_go/lib/app/theme/repository/local_theme_repository.dart
@@ -1,4 +1,4 @@
-import "../app/theme/theme_mode.dart";
+import "../theme_mode.dart";
 
 abstract class LocalThemeRepository {
   Future<void> saveThemeMode(ThemeMode themeMode);
