+++ where_to_go/pubspec.yaml
@@ -32,29 +32,30 @@ dependencies:
   flutter:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  # UI
   cupertino_icons: ^1.0.8
-
-  hooks_riverpod: ^2.6.1
-  flutter_hooks: ^0.21.3
-  riverpod_annotation: ^2.6.1
-  go_router: ^16.1.0
-  drift: ^2.28.0
-  drift_flutter: ^0.2.5
-  path_provider: ^2.1.5
-
-  retrofit: ^4.6.0
-
-  shared_preferences: ^2.5.3
-  flutter_secure_storage: ^9.2.4
-  json_annotation: ^4.9.0
-  dio: ^5.9.0
-  http_parser: ^4.1.2
-  pretty_dio_logger: ^1.3.1
   reactive_forms: ^18.1.1
   image_picker: ^1.2.0
-  reactive_image_picker: ^8.0.1
+
+  # State management
+  hooks_riverpod: ^2.6.1
+  riverpod_annotation: ^2.6.1
+  flutter_hooks: ^0.21.3
+
+  # Navigation
+  go_router: ^16.1.0
+  path_provider: ^2.1.5
+
+  # API
+  dio: ^5.9.0
+  retrofit: ^4.6.0
+  http_parser: ^4.1.2
+  json_annotation: ^4.9.0
+  pretty_dio_logger: ^1.3.1
+
+  # Persistence
+  shared_preferences: ^2.5.3
+  flutter_secure_storage: ^9.2.4
 
 dev_dependencies:
   flutter_test:
@@ -70,7 +71,6 @@ dev_dependencies:
   build_runner: ^2.5.4
   flutter_gen_runner: ^5.11.0
   riverpod_generator: ^2.6.5
-  drift_dev: ^2.28.0
   retrofit_generator: ^9.7.0
   json_serializable: ^6.9.5
